refactor: 作品共通語彙を release パッケージへ抽出し ReleaseNumber へ改名する(#555)
