+++ src/test/kotlin/com/example/api/domain/sakamichi/model/release/PositionTest.kt
@@ -1,4 +1,4 @@
-package com.example.api.domain.sakamichi.model.single
+package com.example.api.domain.sakamichi.model.release
 
 import com.github.michaelbull.result.getError
 import com.github.michaelbull.result.unwrap
