+++ src/test/kotlin/com/example/api/domain/sakamichi/model/single/SingleTest.kt
@@ -3,6 +3,10 @@ package com.example.api.domain.sakamichi.model.single
 import com.example.api.domain.sakamichi.model.group.Group
 import com.example.api.domain.sakamichi.model.group.GroupName
 import com.example.api.domain.sakamichi.model.member.MemberId
+import com.example.api.domain.sakamichi.model.release.Position
+import com.example.api.domain.sakamichi.model.release.ReleaseNumber
+import com.example.api.domain.sakamichi.model.release.Senbatsu
+import com.example.api.domain.sakamichi.model.release.SenbatsuSlot
 import com.example.api.domain.shared.generateId
 import com.github.michaelbull.result.unwrap
 import org.junit.jupiter.api.Test
@@ -10,7 +14,7 @@ import org.junit.jupiter.api.Test
 /** Single 集約の生成のユニットテスト。 */
 class SingleTest {
     private val group = Group.create(GroupName.create("乃木坂46").unwrap())
-    private val number = SingleNumber.create(1).unwrap()
+    private val number = ReleaseNumber.create(1).unwrap()
     private val title = SingleTitle.create("ぐるぐるカーテン").unwrap()
     private val senbatsu =
         Senbatsu.create(listOf(SenbatsuSlot(Position.Center, MemberId(generateId())))).unwrap()
